Move multi-selection infrastructure out of mock vim host
